+++ infra/bootstrapper/_modules/azure/providers.tf
@@ -1,3 +1,11 @@
+terraform {
+  required_providers {
+    dx = {
+      source = "pagopa-dx/azure"
+    }
+  }
+}
+
 provider "github" {
   owner = "pagopa"
 }